refactor!: reorganize pool storage with better structs and better uints
- define 2 small constants as uint16 from uint256, pool implementation constants are initialized in 1 single slot
- removed unitary value and decimals defaults initialization in pool implementation immutables
- mixin storage polish
- return new pool structs in mixin pool state
- return unitary value as either proxy storage or 1 in pool decimals (which are always initialized)
- reorganized pool structs, renamed IStructs as conflicting with staking IStructs
- added natspec docs to IPoolStructs
- added 1 extra struct which is used to return pool proxy initialization parameters

+++ contracts/protocol/core/immutable/MixinConstants.sol
@@ -26,8 +26,7 @@ abstract contract MixinConstants is IRigoblockV3Pool {
     /// @inheritdoc IRigoblockV3PoolImmutable
     string public constant override VERSION = "HF 3.1.0";
 
-    // TODO: we could probably reduce deploy size by declaring smaller constants as uint32
-    uint256 internal constant FEE_BASE = 10000;
+    uint16 internal constant FEE_BASE = 10000;
 
     uint16 internal constant INITIAL_SPREAD = 500; // +-5%, in basis points
 
@@ -36,7 +35,7 @@ abstract contract MixinConstants is IRigoblockV3Pool {
     uint16 internal constant MAX_TRANSACTION_FEE = 100; // maximum 1%
 
     // minimum order size 1/1000th of base to avoid dust clogging things up
-    uint256 internal constant MINIMUM_ORDER_DIVISOR = 1e3;
+    uint16 internal constant MINIMUM_ORDER_DIVISOR = 1e3;
 
     uint16 internal constant SPREAD_BASE = 10000;
